backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
]

[tool.uv]
dev-dependencies = []
